+++ cli/default-reporter/src/reporterForClient/reportSummary.ts
@@ -89,7 +89,7 @@ export function reportSummary (
         if (ignoredScripts.packageNames && ignoredScripts.packageNames.length > 0 && !opts.pnpmConfig?.strictDepBuilds) {
           msg += EOL
           msg += boxen(`Ignored build scripts: ${Array.from(ignoredScripts.packageNames).sort().join(', ')}.
-Run "pnpm approve-builds" to pick which dependencies should be allowed to run scripts.`, {
+Run "pnpm approve-builds${opts.pnpmConfig?.cliOptions?.global ? ' -g' : ''}" to pick which dependencies should be allowed to run scripts.`, {
             title: 'Warning',
             padding: 1,
             margin: 0,