+++ config/config/src/index.ts
@@ -148,6 +148,7 @@ export async function getConfig (opts: {
     'ignore-workspace-cycles': false,
     'ignore-workspace-root-check': false,
     'optimistic-repeat-install': false,
+    'init-package-manager': true,
     'inject-workspace-packages': false,
     'link-workspace-packages': false,
     'lockfile-include-tarball-url': false,