+++ config/plugin-commands-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/plugin-commands-config
 
+## 1000.0.9
+
+### Patch Changes
+
+- @pnpm/cli-utils@1000.0.9
+
 ## 1000.0.8
 
 ### Patch Changes