+++ tools/plugin-commands-self-updater/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/tools.plugin-commands-self-updater
 
+## 1000.0.11
+
+### Patch Changes
+
+- Updated dependencies [91d46ee]
+  - @pnpm/plugin-commands-installation@1001.5.0
+  - @pnpm/cli-utils@1000.0.9
+
 ## 1000.0.10
 
 ### Patch Changes